requirements.txt
openpyxl>=3.1.2
openai>=1.12.0
python-dotenv>=1.0.0
pandas>=2.2.0
numpy>=1.26.0
